Refactor namespace from FrameWriter to DataFrameWriter to follow naming convention
