+++ OpenEphys.Onix1/Bno055DataFrame.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Numerics;
 using System.Runtime.InteropServices;
-using OpenEphys.Onix1.FrameWriter;
+using OpenEphys.Onix1.DataFrameWriter;
 
 namespace OpenEphys.Onix1
 {
@@ -78,7 +78,7 @@ namespace OpenEphys.Onix1
         /// <summary>
         /// Gets the 3D orientation represented as a Quaternion.
         /// </summary>
-        [FrameWriterIgnoreMembers(MemberType.Properties)]
+        [DataFrameWriterIgnoreMembers(MemberType.Properties)]
         public Quaternion Quaternion { get; }
 
         /// <summary>
