+++ OpenEphys.Onix1/DataFrameWriter/ArrowBatchWriter.cs
@@ -4,7 +4,7 @@ using System.Reactive.Linq;
 using System.Reactive.Subjects;
 using Apache.Arrow;
 
-namespace OpenEphys.Onix1.FrameWriter
+namespace OpenEphys.Onix1.DataFrameWriter
 {
     /// <summary>
     /// Provides buffered writing of items to an Arrow file, batching items before writing.
