+++ OpenEphys.Onix1/DataFrameWriter/IRecordBatchExpressionProvider.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Reflection;
 
-namespace OpenEphys.Onix1.FrameWriter
+namespace OpenEphys.Onix1.DataFrameWriter
 {
     interface IRecordBatchExpressionProvider
     {
